restructued selectors into specfici files that reference their own reducers and updated tests files (passing).
