+++ src/selectors/userSelectors.js
@@ -1,0 +1,1 @@
+export const getUserMixtapes = state => state.user.mixtapes;
